refactor 클래스 통합
