+++ src/main/java/lotto/model/LottoShop.java
@@ -1,5 +1,10 @@
 package lotto.model;
 
+import java.util.ArrayList;
+import java.util.List;
+import lotto.model.generator.LottoGenerator;
+import lotto.model.generator.LottoNumberGenerator;
+
 public class LottoShop {
     public static final int LOTTO_PRICE = 1000;
 
@@ -7,4 +12,14 @@ public class LottoShop {
         return money / LOTTO_PRICE;
     }
 
+    public LottoWallet getLottoNumbers(int size) {
+        LottoGenerator lottoGenerator = new LottoGenerator(new LottoNumberGenerator());
+
+        List<Lotto> lottos = new ArrayList<>();
+        for(int i = 0; i < size; i++) {
+            lottos.add(new Lotto(lottoGenerator.generate()));
+        }
+
+        return new LottoWallet(lottos);
+    }
 }
